// src/components/Badge/Badge.types.ts
export type BadgeVariant =
  | 'default'
  | 'secondary'
  | 'outline'
  | 'outline-success'
  | 'destructive'
  | 'warning'
  | 'green'
  | 'lime'
  | 'emerald'
  | 'teal'
  | 'cyan'
  | 'cyan-light'
  | 'sky'
  | 'blue'
  | 'indigo'
  | 'violet'
  | 'purple'
  | 'fuchsia'
  | 'pink'
  | 'rose'
  | 'navy'
  | 'gray'
  | 'neutral-light'
  | 'white'
  | 'green-light'
  | 'none'
  | 'vivid-purple'
  | 'bright-blue'
  | 'rose-red';